+++ CHANGELOG.md
@@ -6,4 +6,5 @@
 - Add support for keys other than `4K`.
 - Allow (maybe configurable) filters.
 - Remove input override in order to streamline config.
-- Add proper logging.
+- Add a directory junction mode.
+- Use difficulties from `osu!.db`.